+Añadiendo modificacion de solicitudes con elementos O2M

+++ fap/compiler/src/es.fap.simpleled.generator/src/generator/utils/Controller.groovy
@@ -1,5 +1,7 @@
 package generator.utils;
 
+import java.util.ArrayList;
+
 import es.fap.simpleled.led.Accion
 import es.fap.simpleled.led.Attribute;
 import es.fap.simpleled.led.Boton
@@ -34,6 +36,7 @@ import org.eclipse.emf.ecore.EObject
 import org.eclipse.emf.ecore.resource.Resource
 import org.eclipse.xtext.generator.IFileSystemAccess
 import org.eclipse.xtext.generator.IGenerator
+
 
 import templates.GBoton;
 import templates.GElement
@@ -586,6 +589,9 @@ public class ${controllerName} extends ${controllerGenName} {
 	}
 	
 	private String metodoBorrar(){
+		String codigoBorrar = "";
+		String backupCopia = "";
+		
 		if (!borrar)
 			return "";
 		def borrarEntidad = "";
@@ -596,11 +602,43 @@ public class ${controllerName} extends ${controllerGenName} {
 				borrarEntidad += "db${lastSubcampo.str} = null;\n";
 			borrarEntidad += "${almacen.variableDb}.save();\n";
 		}
-		if (!noBorrarEntidad) {
+		if ((!noBorrarEntidad) && (!copia)){
 			borrarEntidad += """
 				${borrarListasXToMany()}
 				${entidad.variableDb}.delete();
 			""";
+		}
+		if (copia){
+			backupCopia = """
+				PeticionModificacion peticionModificacion = new PeticionModificacion();
+				peticionModificacion.campoPagina="${campo.str}";
+				
+				Map<String, String> allSimple = params.allSimple();
+				for(Map.Entry<String, String> entry : allSimple.entrySet()){
+					if(entry.getKey().startsWith("id")){
+						 try {
+						    peticionModificacion.idSimples.put(entry.getKey(), Long.parseLong(entry.getValue()));
+						 }catch(Exception e){
+						//El parámetro no era un long
+					}
+				}
+	         }
+			List<String> valoresAntiguos = new ArrayList<String>();
+			boolean hayModificaciones = false;
+			"""			
+			campo.ultimaEntidad.attributes.eachWithIndex { item, i ->
+			if (!item.name.startsWith("id"))
+				codigoBorrar += """
+				if (${entidad.variableDb}.${item.name} != null){
+					valoresAntiguos.add(${entidad.variableDb}.${item.name}.toString());
+					peticionModificacion.setValorBorrado("${entidad.variableDb}.${item.name}", new ArrayList<String>(), valoresAntiguos);
+					hayModificaciones = true;
+				}
+			"""
+			}
+			codigoBorrar += """
+							${gElement.saveCode()}
+							"""
 		}
 		return """
 			public static void borrar(${StringUtils.params(allEntities.collect{it.typeId})}){
@@ -613,6 +651,12 @@ public class ${controllerName} extends ${controllerGenName} {
 				if(!Messages.hasErrors()){
 					${controllerName}.borrarValidateRules(${StringUtils.params(entidad.variableDb)});
 				}
+				
+				if (!Messages.hasErrors()) {
+					${backupCopia}
+					${codigoBorrar}
+				}
+
 				if(!Messages.hasErrors()){
 					$borrarEntidad
 					log.info("Acción Borrar de página: "+${renderView}+" , intentada con éxito");
@@ -920,23 +964,65 @@ public class ${controllerName} extends ${controllerGenName} {
 	public String metodoValidateCopy(){
 		if (saveEntities.size() == 0 || (!editar && !crear)) return "";
 		String backupCopia = "";
-		if (copia)
-			backupCopia = """PeticionModificacion peticionModificacion = new PeticionModificacion();
-							 peticionModificacion.campoPagina="${campo.str}";
-							 boolean hayModificaciones=false;
-							 Map<String, String> allSimple = params.allSimple();
-							 for(Map.Entry<String, String> entry : allSimple.entrySet()){
-								 if(entry.getKey().startsWith("id")){
-									 try {
-									    peticionModificacion.idSimples.put(entry.getKey(), Long.parseLong(entry.getValue()));
-									 }catch(Exception e){
-										//El parámetro no era un long
-									 }
-								  }
-		                     }
-							 List<String> valoresAntiguos = new ArrayList<String>();
-							 List<String> valoresNuevos = new ArrayList<String>();
-			"""
+		String copiaTexto = "";
+		String codigoCopia="";
+		String cierre = "";
+		String codigoCrear="";
+		
+		if (copia){
+			backupCopia = """
+				PeticionModificacion peticionModificacion = new PeticionModificacion();
+				peticionModificacion.campoPagina="${campo.str}";
+				boolean hayModificaciones=false;
+				Map<String, String> allSimple = params.allSimple();
+				for(Map.Entry<String, String> entry : allSimple.entrySet()){
+					if(entry.getKey().startsWith("id")){
+						 try {
+						    peticionModificacion.idSimples.put(entry.getKey(), Long.parseLong(entry.getValue()));
+						 }catch(Exception e){
+						//El parámetro no era un long
+					}
+				}
+	         }
+			List<String> valoresAntiguos = new ArrayList<String>();
+			List<String> valoresNuevos = new ArrayList<String>();
+			"""				
+		}
+		
+		//Validate copy de pagina de tabla
+		if ((copia) && (!saveEntities.collect{ it.variableDb }.contains("dbSolicitud"))){
+			copiaTexto = """
+				String idSolicitud = params.get("idSolicitud");
+				SolicitudGenerica dbSolicitud = SolicitudGenerica.findById(Long.parseLong(idSolicitud));
+				Boolean creando = false;
+			""";
+			
+			//Comprobacion de nuevo elemento en modificacionSolicitudes
+			List<String> lista = saveEntities.collect{"${it.typeVariable}"}.join(" ").toString().split(" ");
+			codigoCopia = """if (db${lista.get(0)}.id == null){
+						db${lista.get(0)}.save();
+						params.put("id${lista.get(0)}", db${lista.get(0)}.id.toString());
+						creando = true;
+						}
+						""";
+			List<String> atributos =  new ArrayList<String>();
+			codigoCrear = """if (creando) {"""
+			campo.ultimaEntidad.attributes.eachWithIndex { item, i ->
+				atributos.addAll(item.name);
+				if (!item.name.startsWith("id"))
+					codigoCrear += """
+					if (${lista.get(1)}.${item.name} != null){
+						valoresNuevos = new ArrayList<String>();
+						valoresNuevos.add(${lista.get(1)}.${item.name}.toString());
+						peticionModificacion.setValorCreado("${lista.get(1)}.${item.name}", new ArrayList<String>(), valoresNuevos);
+						hayModificaciones = true;
+						db${lista.get(0)}.${item.name} = ${lista.get(1)}.${item.name};
+					}
+					"""
+			}
+			codigoCrear += "}";
+		}
+		
 		return """
 			@Util
 			public static void ${name}ValidateCopy(String accion, ${StringUtils.params(
@@ -947,7 +1033,10 @@ public class ${controllerName} extends ${controllerGenName} {
 				extraParams
 			)}){
 				CustomValidation.clearValidadas();
+				${copiaTexto}
+				${codigoCopia}	
 				${backupCopia}
+				${codigoCrear}
 				${gElement.validateCopy()}
 				${gElement.saveCode()}
 			}
@@ -1592,5 +1681,9 @@ public class ${controllerName} extends ${controllerGenName} {
 		else
 			return"";
 	}
-			
+	public Collection getItvariableDb(){
+		return saveEntities.collect{it.variableDb};
+	}
+	
+				
 }
